requirements.txt
asgiref==3.2.10
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coinbase==2.1.0
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
enum34==1.1.10
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
oauthlib==3.1.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.51.0
tornado==6.0.4
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
whitenoise==5.2.0
zope.interface==5.1.0